+++ CHANGELOG.md
@@ -12,6 +12,7 @@
         * `explanationEntitiesField`
         * `openPeriodField`
         * `closeDateField`
+    * Field `SendPoll#closeInfo` was added
     * `SendQuizPoll` now able to use fields `caption` and `parseMode` for `explanation` functionality
         * `quizPollExplanationLimit` was added for checking `QuizPoll` explanation size
     * Field `TextLinkTextSource#url` was added